% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFeature.R
\name{getFeature}
\alias{getFeature}
\title{Get genomic ranges based on specified features and flanking regions.}
\usage{
getFeature(
  object,
  start_feature = "TSS",
  start_flank = 0,
  start_exon = NULL,
  start_exon_boundary = NULL,
  start_direction = "up",
  end_feature = "TES",
  end_flank = 0,
  end_direction = "down",
  end_exon = NULL,
  end_exon_boundary = NULL,
  return = "all"
)
}
\arguments{
\item{object}{A \code{GRanges} object containing genomic features (e.g., genes
with exons). This object must have metadata columns including 'blockStarts',
'blockSizes', and 'blockCount' if 'Exon' is specified as a feature.}

\item{start_feature}{A character string specifying the starting feature.
Must be one of "TSS" (Transcription Start Site), "TES" (Transcription
End Site), or "Exon". Default is "TSS".}

\item{start_flank}{An integer indicating the number of bases to flank the
\code{start_feature}. Default is 0.}

\item{start_exon}{An integer specifying the exon number to use if
\code{start_feature} is "Exon".}

\item{start_exon_boundary}{A character string specifying the exon boundary to
use if \code{start_feature} is "Exon". Must be either "5prime" or
"3prime".}

\item{start_direction}{A character string indicating the direction of the
flank relative to the \code{start_feature}. Must be either "up"
(upstream) or "down" (downstream). Default is "up".}

\item{end_feature}{A character string specifying the ending feature.
Must be one of "TSS", "TES", or "Exon". Default is "TES".}

\item{end_flank}{An integer indicating the number of bases to flank the
\code{end_feature}. Default is 0.}

\item{end_direction}{A character string indicating the direction of the
flank relative to the \code{end_feature}. Must be either "up"
(upstream) or "down" (downstream). Default is "down".}

\item{end_exon}{An integer specifying the exon number to use if
\code{end_feature} is "Exon".}

\item{end_exon_boundary}{A character string specifying the exon boundary to
use if \code{end_feature} is "Exon". Must be either "5prime" or
"3prime".}

\item{return}{A character string specifying which ranges to return. Must be
one of "all" (both plus and minus strand ranges), "plus" (only plus
strand ranges), or "minus" (only minus strand ranges). Default is "all".}
}
\value{
A \code{GRanges} object containing the extracted or created genomic ranges.
}
\description{
This function takes a GRanges object and extracts or creates new genomic
ranges based on specified start and end features (TSS, TES, or Exon) and
their flanking regions. It allows for specifying exon numbers and boundaries
for precise range definition.
}
\examples{
# Create a dummy GRanges object
library(GenomicRanges)
genes <- GRanges(
  seqnames = "chr1",
  ranges = IRanges(start = c(100, 2000), end = c(1000, 3000)),
  strand = c("+", "-"),
  blockStarts = c("0,200,500", "0,300"),
  blockSizes = c("100,300,500", "200,500"),
  blockCount = c(3, 2)
)

# Get TSS regions with 100bp upstream flank
tss_regions <- getFeature(genes, start_feature = "TSS", start_flank = 100)
tss_regions

# Get regions from TSS to TES
gene_bodies <- getFeature(genes, start_feature = "TSS", end_feature = "TES")
gene_bodies

# Get regions around the 2nd exon start
exon2_start <- getFeature(genes, start_feature = "Exon", start_exon = 2,
                           start_exon_boundary = "5prime", start_flank = 50)
exon2_start

# Get regions spanning the 1st to the 2nd exon
exon1_to_2 <- getFeature(genes, start_feature = "Exon", start_exon = 1,
                          start_exon_boundary = "5prime",
                          end_feature = "Exon", end_exon = 2,
                          end_exon_boundary = "3prime")
exon1_to_2

}
